Consolidate notice entrypoints under entrypoints/notices/
Per Mayisha's PR review: move the three sibling notice entrypoints
(one-and-done, test-to-live, post-kyc-activation) under
client/entrypoints/notices/ and replace the duplicate per-entrypoint
style.scss files with a single shared one. The shared rule covers all
three banner IDs.

Updates webpack/shared.js to the new paths and the docblock
breadcrumbs in class-wc-payments-admin-banner.php.

No behavior change.
