+++ client/entrypoints/notices/one-and-done-notice/components/notice/index.tsx
