+++ client/entrypoints/notices/post-kyc-activation-notice/components/notice/index.tsx
