--- client/entrypoints/test-to-live-notice/style.scss
@@ -1,12 +1,0 @@
-// On WC Admin SPA pages (Overview, Marketing, Home, Analytics, …) the page
-// content is centered in a narrow column. Constrain the banner to match so
-// it doesn't span dramatically wider than the content it sits above.
-// We detect SPA pages via the presence of .woocommerce-layout__main — the
-// `body.woocommerce-admin-page` class is too broad (it's also added on
-// embed-mode classic pages like Settings, where the banner should stay
-// wide to match the settings form below). 680px matches
-// .woocommerce-payments-page.is-narrow (client/components/page/style.scss).
-body:has( .woocommerce-layout__main ) #wcpay-test-to-live-notice {
-	max-width: 680px;
-	margin: 0 auto;
-}
